pyproject.toml
[tool.poetry]
name = "habits-aiogram"
version = "0.1.0"
description = ""
authors = ["Mikhail Sobolev <kseon@inbox.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.12"
aiofiles = "23.2.1"
aiogram = "3.3.0"
aiohttp = "3.9.5"
aiosignal = "1.3.1"
aiosqlite = "0.19.0"
annotated-types = "0.7.0"
attrs = "23.2.0"
certifi = "2024.6.2"
frozenlist = "1.4.1"
greenlet = "3.0.3"
idna = "3.7"
magic-filter = "1.0.12"
multidict = "6.0.5"
porty = "0.0.1"
pydantic = "2.5.3"
pydantic-core = "2.14.6"
python-dotenv = "1.0.1"
sqlalchemy = "2.0.27"
typing-extensions = "4.12.2"
yarl = "1.9.4"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
